// frontend/src/openapi/models/FinetuningExcludedModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface FinetuningExcludedModel
 */
export interface FinetuningExcludedModel {
    /**
     *
     * @type {string}
     * @memberof FinetuningExcludedModel
     */
    model: string;
    /**
     *
     * @type {string}
     * @memberof FinetuningExcludedModel
     */
    reason: string;
}

/**
 * Check if a given object implements the FinetuningExcludedModel interface.
 */
export function instanceOfFinetuningExcludedModel(value: object): value is FinetuningExcludedModel {
    if (!('model' in value) || value['model'] === undefined) return false;
    if (!('reason' in value) || value['reason'] === undefined) return false;
    return true;
}

export function FinetuningExcludedModelFromJSON(json: any): FinetuningExcludedModel {
    return FinetuningExcludedModelFromJSONTyped(json, false);
}

export function FinetuningExcludedModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): FinetuningExcludedModel {
    if (json == null) {
        return json;
    }
    return {

        'model': json['model'],
        'reason': json['reason'],
    };
}

export function FinetuningExcludedModelToJSON(json: any): FinetuningExcludedModel {
    return FinetuningExcludedModelToJSONTyped(json, false);
}

export function FinetuningExcludedModelToJSONTyped(value?: FinetuningExcludedModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'model': value['model'],
        'reason': value['reason'],
    };
}